Add GitHub Actions workflow to publish NuGet package
This commit introduces a new GitHub Actions workflow in `dotnet.yml`
to automate publishing a package to NuGet. The workflow is triggered
on `push` events to the `master` branch and manual `workflow_dispatch`
events. It is scoped to changes in the `src/VmmSharpEx/**` directory.

A concurrency group is defined to ensure only one workflow run per
branch is active at a time, canceling any in-progress runs if a new
one is triggered. The `build-and-publish` job runs on `ubuntu-latest`
and uses the `production` environment for deployment.